知识点开发

+++ src/main/java/com/fs/ntes/domain/Question.java
@@ -19,7 +19,7 @@ public class Question {
 
     private String analysis;
 
-    private Integer uid;
+    private Long uid;
 
     private Integer status;
 
@@ -91,11 +91,11 @@ public class Question {
         this.analysis = analysis == null ? null : analysis.trim();
     }
 
-    public Integer getUid() {
+    public Long getUid() {
         return uid;
     }
 
-    public void setUid(Integer uid) {
+    public void setUid(Long uid) {
         this.uid = uid;
     }
